Platform agnostic keybinding
Currently, if a keybinding is defined for Cmd-S and the user types
Ctrl-S, the event is not dispatched to the handler. Likewise, keybindings
defined with Ctrl are not handled if the user types Cmd. Platform
agnostic keybinding automatically listens for both the Cmd and Ctrl
variants of a keybinding.

Platform agnostic keybinding is opt-in, requiring the component to
declare it's intention with a property in the form:

    keybindingsPlatformAgnostic: true,
    keybindings: { ... }

Issue: #7